One word fix for slightly awkward phrasing
Right now when running e.g. `wait4x exec true`, the output shows "Checking the true", which is awkward grammatically.

"Checking true" would be slightly better, or even better would be some message which is specific to the kind of condition being waited on ("Checking execution of true") but the latter seems more invasive of a change, so perhaps just dropping "the" is acceptable?